Merge branch dev in branch main

+++ admin/gallery/delete.php
@@ -4,20 +4,25 @@
     use Src\Models\Gallery;
 
     $gallery = new Gallery();
+    $message = 'Image(s) removida(s) com sucesso!';
+    $type = 'success';
 
     foreach(requests()->ids as $id):
         $image = $gallery->find($id);
 
-        isset($image->data) && deleteDir(__DIR__."/../../public/assets/images/{$image->data->file}");
-
-        $image->posts()->detach($id);
-
-        $image->delete();
+        if($image->categories()->data):
+            $message = 'Algumas imagens não foram removidas por estarem vinculadas a alguma categoria!';
+            $type = 'info';
+        else:
+            $image->locations()->detach($id);
+            $image->delete();
+            isset($image->data) && deleteDir(__DIR__."/../../public/assets/images/{$image->data->file}");
+        endif;
     endforeach;
 
     session([
-        'message' => 'Image(s) removida(s) com sucesso!',
-        'type' => 'success'
+        'message' => $message,
+        'type' => $type
     ]);
 
     return header(route('/admin/gallery', true), true, 302);
